requirements.txt
pillow
segno
qrcode-artistic
click
